python/requirements.txt
Flask>=2.0
psycopg2-binary>=2.9
redis>=4.0
pika>=1.2.0